Resolving conflict: Updated a pipe in simple flow tutorial
While rebasing, the original commit message became obsolete. There used
to be an additional (unnecessary) pipe in the simple flow tutorial.
Since the commit that added the additional pipe was removed from the
ancestor history, there is nothing to delete now.

However, the other of the two pipes was modified in an even earlier
commit by @jschnaitmann. The reason why this was done is unknown however
it seems innocuous and justifiable enough to keep the change going
forward.

Was: Delete an unnecessary pipe from the simple flow tutorial