Fixup calculation of total molar mass
The pressure calculation in the phase update was totally broken due to
miscalculation of the total molar mass. The |calculateMolecularMass|
method of the matter table used to calculate some sort of mean molar
mass based on the mass fraction. However, it should be based on the
amount of substances in the mixture. This commit changes the way the
total molar mass is calculated and documents it.