Change density calculation of pipe component
When calculating the solver deltas, the pipe used to only take the
density of the incoming flow into account. This is assumed to be
inaccurate. As a first measure, this commit changes the assumed density
to be the arithmetic mean of the densities of both the incoming and
outgoing flow.

This change is yet to be validated.