Make sure to return correct EXME for incoming flow in branch
When the flow rate is zero, the |getInEXME| method of the branch class
may have returned the wrong EXME. In this case, the method now returns
the EXME that is connected to the phase with the higher mass.

Note that this change may still lead to the wrong EXME being reported
but it ensures that the matter properties do not become zero if the
first phase is empty.

Code by @jschnaitmann