Timer: execute more callbacks by also including those in the "near future"
When the timer executes, it determines which callbacks to execute at the current time.
It also includes callbacks in the "near future".

This behavior is extended: in the past, only callbacks within 1/2 of the current time step
in the future were executed, now nearly a full time step is checked. Should help to "sync"
e.g. two phases with large time steps so they execute at the same tick.