Add adaptive rMaxChange in Phases
A seconds parameter, rHighestMaxChangeDecrease, can be set. If it is set, the lower the changes
in the phase contents, the more rMaxChange will be decreased. Depending on the "magnitude" of
the mass change (i.e. change in percent per second - magnitude is the "zeros behind the dot"),
rMaxChange will be multiplied with a factor interpolated between 1 and rHighestMaxChangeDecrease.

A global parameter on oData (same name) can be set and will be used by all phases.

rHighestMaxChangeDecrease is zero by default, i.e. should introduce no new behaviour.