First version of the residual solver
Can be used once for a phase, solver should be added as the last solver.
Sums up flowrates of branches (including p2ps) and sets this value as flow rate.

Note: only checks/works for the "left" phase, has to be added accordingly in .createBranch().
Will not be 100% exact due to rounding errors (at least that is the excuse).