Fix |simulation.masslog| to sum over correct columns
The |reshape| function was set up to re-order the rows as columns,
this lead to unrelated substance masses being summed. The fix includes
reordering the parameters to the |reshape| function and transforming
the sum vector. The additional parameter to |sum| makes sure the
correct dimension is summed.

This bug did not have an influence on the total mass balance but could
nonetheless confuse users when debugging systems.