Only run changed tutorials in |runAllTutorials|
When |runAllTutorials| is executed, only tutorials whose files have
changed since the last execution of this function will be re-run. If any
files were changed in /core/ or /lib/, all tutorials will be run. If no
files were changed, nothing is run.

The function saves the actual change date in a MAT file in the /data/
folder. It also checks for the files in the root folder.