Align solid phase with other phases
The gas and liquid phase classes take a volume as the fourth parameter
to their constructor. The solid class, however, has not, which is
unexpected and leads to incompatibilities with some aspects of the
thermal system. This commit adds handling volume to the phase, however
it is basically ignored and calculated from the material's density
(since solid phases are assumed to be incompressible).

In the same spirit, an |fPressure| property has been added that defaults
to |NaN| so |calculateHeatCapacity| and friends does not throw an error
(the standard pressure should be used in this case).