p2p getInFlows - remove the partials manipulator flow rate
Previously, the partial masses "produced" by the manipulator were added to the total in-flowrates.
This does not make sense - for once, manipulators should destroy as much as they produce, i.e. do
not really add the the total flow rate. Also, manipulators are executed AFTER p2ps, and the more
likely influence is p2p -> manipulator, not vice versa.