BREAKING: Massive overhaul of matter table
The matter data is now stored in csv files that are more transparent for git in terms of tracking changes. This causes major changes to the importMatterData() method of matter.table.
All files associated with the Excel file handling were deleted.
As far as possible, all core, lib and tutorial files have been adapted to the new matter table.
Next to the changes in the code, this commit includes the actual data files.
The matter table now discerns between isochoric and isobaric data. This is very important, since V-HABs phases and stores all have fixed volumes, so we should have been using isochoric values all along. Unfortunately, we used isobaric values, so there should be some changes to the simulation results, especially with respect to heat capacities.
The API for the |findProperty()| method was changed. Instead of entering 8+ parameters in a certain order, the method now accepts a struct with certain field names, some of which are optional. Due to its complexity, the |findProperty()| method was also moved out of the table.m file into the @table folder.