Removed cause for warning message in arrangeWindows() when using MATLAB 2015b or later.
The ScreenPixelsPerInch property of the graphics root (groot) object was made read-only in 2015b. It was being used to boost the font size on Macs because the default value of 72 ppi was way too low for modern displays.
Now the font size is increased via the FontSize property of the axes object.