Fixed an error in the volumetric flow rate calculation in flow.m
Previously, the flow rate was calculated using the flow rate from the exme on the inflow side of the branch. This is however not changed during the iterations of the iterative solver, so the flow rate used by the function would always be the same, thus it would always return the same volumetric flow rate, even if the actual flow rate was varied by the solver. This impacted the delta pressure calculations in the fan model.

Now the calculateVolumetricFlowRate() method has an optional input parameter for the flow rate.

The fan component has also been changed accordingly.