Added some error handling to the logger.
An error would not be detected if the user tried to log values that have not been set yet and are 'empty'. This would result in all log items being discarded and the log showing NaN for all values.

The expanded error handling catches this specific situation and replaces the 'empty' return value with 'NaN', keeping the other log item values intact.