Added checking the matter table itself for changes at the beginning of the constructor.
Previously, if the matter table itself or any of the functions contained in the @-folder had changed, it would not have triggered a reconstruction of the matter table object. Instead the object saved in the MatterData.mat file would have been loaded and this may cause errors because of mission properties and methods.
Now if any file in the @table folder has changed, the matter table object will be reconstructed.