implemented a fix for the residual solver, as for what it does:
Assume Residual Solver A is attached to Phase A and B and is supposed to keep the mass of Phase A constant. If then Residual Solver B is attached to Phase B and C and is supposed to keep the mass of Phase B constant then residual solver B is required to update in case that residual solver A changes its flowrate. The issue here is that obviously all residual solvers are executed with the same post tick prio. The solution was to implement a check if any other residual solvers are attached to the same outlet phase as the current residual solver. So while residual solver A keeps the mass of Phase A constant, it checks if any residual solvers are connected to Phase B and orders a new posttick update for all residual solvers attached to Phase B. Note that it would be preferable that only the residual solvers that are supposed to keep the mass in phase B constant are reupdated (as there should be only one) but ensuring that this remains the case even if the other solver changes the positive flow direction of the residual is fairly difficult and left for future work ;)