cherry pick of commit: ba0e4e6dc036f93a4659d4ba90668b5e89511ecb [ba0e4e6] from @olcl
Original Commit Message:
In plotter_basic created new definePlot() and defineFigure() methods. Old version of the former is still included for reference as definePlot_old().
Added a adjustTime() method to enable units of time other than seconds.
Added a method that adds another plot with the y axis on the right.
Added a method to parse graphics object properties from a user-provided struct and set them on a given graphics object.
In logger_basic added a method to calculate the number of individual units in one plot.