Move new bSynced behaviour behind other flag, include complete functionality
In an earlier commit, the meaning of bSynced was extended so that when set,
the connected EXMEs would use the inflowing matter composition instead of
the one within the phase.

This was somewhat reverted now, and instead, this new behaviour was moved
behind a flag bFlow, as it was realized that some simulations seem to use
bSynced in situations where the bFlow behaviour is probably not intended.

Additionally, the functionality was finalized with some long outstanding,
uncommited local changes to include all relevant matter properties, procs,
and some more comments.