CORE: branch names are normalized to allow branch names in the GUI to have spaces
FEATURE: Added import functionality for xml files created with the Draw.IO V-HAB Library (which is in the examples DrawIO folder) as well as two examples of how this can be used. Is currently missing documentation, which is still a to do and should be done in the wiki and the code.