CORE: There was an error in the connectTo() method in the base branch that caused thermal capacities on the subsystem and supersystem to not be correctly connected. This was because of the differences between matter and thermal branches. The thermal branch needs to be set on the thermal exme 'oBranch' property. This was not being done. This in turn led to the thermal branch on the supersystem not being assigned a thermal solver branch and thus never influencing the temperature of the capacity it is connected to.
This commit adds the setting of oBranch on the thermal exme in the connectTo() method of the base branch. Since the addBranch() method of the thermal exme checks for an existing branch, this method had to edited as well to enable the setting of the branch only if it is the correct branch from the correct subsystem that is being set.