CORE: In an effort to get rid of the "Objects of class X still exist and cannot be cleared" warnings during initialization, all core classes were checked for objects in the properties. This mostly concerned object arrays, such as aoFlows, being populated with an empty object, here matter.flow.empty(). This is mentioned in the MathWorks forums as one of the causes for the above mentioned warning. When this is not taken care of, the time it takes to clear the classes after a nominal simulation run grows exponentially. The changes made in this commit partially remove the warnings, but not always. Especially when a crash occurs during simulation model assembly, there are apparently lots of orphaned objects. Still need to investigate.
CORE: In an effort to get rid of the "Objects of class X still exist and cannot be cleared" warnings during initialization, all core classes were checked for objects in the properties. This mostly concerned object arrays, such as aoFlows, being populated with an empty object, here matter.flow.empty(). This is mentioned in the MathWorks forums as one of the causes for the above mentioned warning. When this is not taken care of, the time it takes to clear the classes after a nominal simulation run grows exponentially. The changes made in this commit partially remove the warnings, but not always. Especially when a crash occurs during simulation model assembly, there are apparently lots of orphaned objects. Still need to investigate.