CORE: Massive change in the way the thermal domain objects are created. This was done in an effort to better separate the two domains. The trigger for this was the implementation of the thermal multi-branch solver, where capacities need to be of a different type (thermal.capacity.network).
Now, instead of all of the thermal objects being created in parallel within the matter domain, specifically during createMatterStructure(), all thermal objects are created in createThermalStructure(). The exception are the thermal branches associated with the matter branches, these are created separately after createThermalStructure() on the simulation container has completed. The existing code should not perform differently after this change. The only breakage that occurs is with respect to the names of the thermal exmes. There is a lengthy explanation of this problem in thermal.container.createAdvectiveThermalBranches(). As a result, code that relied on the fact that the thermal exmes have the same name as the matter exmes will break. In a following commit, the tests, tutorials, examples and library components will be adapted to this change.

This sequential creation of matter and then thermal also makes small changes necessary regarding where reference object properties are set. (e.g. matter.phase.oCapacity, matter.branch.oThermalBranch, etc.) As part of the automatic generation it needed to be ensured that exme names within a store are unique. Until now, they were only checked to be unique within a phase.

The initial issue of creating thermal network capacities instead of normal capacities, if they are to be used in the thermal multi-branch solver, is solved by creating a property in matter.phase called bThermalNetworkNode. This is false by default and can be set after the phase is created.

Finally, the API of the thermal branch has been changed to accept only thermal capacities as input objects and will throw an error if a phase is passed in.