CORE: Major update to the way the specific heat capacity property is updated in phases and flows. Previously, the value for specific heat capacity was always taken from the phase. This approach created incorrect results, especially when the phase in question is a flow phase. The problem mainly manifested itself in the update() method of the fluidic conductor class. Now the value is calculated based on the matter flows that go into the processor or branch associated with a fluidic conductor. This required many changes to how and when the value is calculated. Also, in both the fluidic conductor and the basic thermal solver, the property oMassBranch was re-named to oMassObject, since it can be a matter branch or a P2P processor.
Since calling calculateSpecificHeatCapacity() on the matter table is fairly expensive computationally and with the changes described above it would be used even more than before, all of the implementations in this commit include checks if the relevant temperatures, pressures and partial masses have changed enough to warrant a recalculation.

During the implementation of the changes it was also noticed, that the setMatterProperties() method of the flow p2p could be moved into the base p2p class. There, some comments where added in addition to the specific heat capacity calculation code.

Somewhat intertwined with the issue regarding the specific heat capacity is the calculation of the afPP property of gas flow phases. This value should not be used while the matter multi-branch solver is iterating, so the property was made dependent and an error is thrown if it is accessed while an update of the solver is in progress. During the implementation of that change it was noticed, that the afPartialPressure property of the matter flow class is also recalculated very often, using a call to calculatePartialPressures() on the matter table, and also rarely used. So it was decided to make it a dependent property as well.