CORE: Small but important change to calculateTimeStep() in matter.phase. When setting the time step for flow phases on the timer, the second argument of setTimeStep() is a true boolean. This causes the timer to reset the last execution time for that phase to the current time. This prevents those flow phases from being "overtaken" by the time step and causing ticks with the minimum time step.
An example is a phase where at T = 5 the time step is set to 20 seconds, scheduling the next update for T= 25. In the following ticks, other objects may trigger a re-calculation of the time step. It could happen, that at T = 22 is calculated as 15 seconds, scheduling the next execution for T = 20, which is obviously in the timer's past. That will cause the fThisStep variable calculated in the tick() method to become negative and the timer will set the minimum time step of 1e-8.