CORE: Bugfix and performance enhancement in the advanced thermal multi-branch solver.
The bug related to the way the matrices of unique and non-unique capacities was built. It failed to account for the case where capacities exist that are purely radiative, meaning they are not connected to other capacities by conductive branches, only by radiative branches.

The performance enhancement was achieved by reducing the size of the matrices that are used in the calculateTemperatureChangeRate() method, which is the method called by the ODE solver in update(). Previously, the entire mbNonUniqueCapacitiesTransposed matrix was used, now only the radiative capacities are involved in the operation that retrieves the temperatures of those nodes. Depending on the size of the network and the number of linear and radiative nodes, this can significantly increase performance.