Bugfixes for Thermal Manual Branch and Thermal Fluidic Branch used in Multibranch matter solvers
Previously the thermal manual could get stuck with the flag that it is outdated and the update would not be run as the requested heat flow and current heat flow beeing identical. This led to the thermal manual branch beeing stuck in this state even if the requested heat flow did change in the future. For the thermal fluidic branch it was possible that only on matter branch requested the update of the multibranch solver, thus not setting the other associated thermal fluidic branches to outdated. This is now caught by having the matter multibranch solver also set all associcated thermal branches to outdated during its register update routine
This required a fix for the vacuum interface pipes of CDRA as during the solver update the specific heat capacity update failed due to unrealistic pressures
Also fixed a bug in two test cases for the case when no simulation time is provided for the test run of V-HAB