Fixed Bugs for Parallel Execution
in the "generalParallelExecution" the nargin properties were not correct, this was fixed

in the multiWaitbar updateing of the bar background can fail, leading to a crash of all calculations. As the background is not that important, instead implemented a try catch which lets the simulations continue while only providing a warning that the update failed in case this happens.